requirements.txt
click==8.1.7
Flask==2.3.3
fpdf==1.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-pptx==0.6.21
six==1.16.0
Werkzeug==2.3.7
flask-ngrok==0.0.25
